backend/requirements.txt
fastapi==0.110.1
uvicorn[standard]==0.29.0
SQLAlchemy==2.0.29
asyncpg==0.29.0
alembic==1.13.1
python-jose==3.3.0
passlib[bcrypt]==1.7.4
pydantic==1.10.14
pydantic-settings==2.2.1
python-multipart==0.0.9
httpx==0.27.0
pytest==8.1.1
pytest-asyncio==0.23.6
anyio==4.3.0
